[ADD] l10n_ar_edi: RG 4919/2021 FCE Transmission
Thanks to new changes in AFIP we need to sent FCE Trasmission option
information for invoice validation when the type of document is mipyme.

We add a new system parameter that let the user to define their default
transmission option and we update it the logic to sent the selected FCE
Transmission option to the AFIP when validating the document.

* RG 69/2020 Update unit tests

 The minimum amount for MiPyME documents has been updated, before was
 100.000 ARS and now is 146.885 ARS. We update the unit tests to be able to
 proper create, validate and tests MiPyME documents using this new
 minimum.

 more information in https://www.afip.gob.ar/noticias/20200703-Emision-de-FCE.asp

closes odoo/enterprise#16540

Signed-off-by: Josse Colpaert <jco@openerp.com>