[FIX] l10n_be_coda: no crash in float parsing if empty value
In some structured communications, we are expecting float values when parsing the
communication (example: volume for transaction type fuel in communication 113). But
this data is not always filled in, and we should not crash in that case but rather
just set a default 0 value to allow the import of CODA statements.

closes odoo/enterprise#16609

X-original-commit: 9308e9937a6394fd7e868a7f2f2c9c2837bc7497
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>