[FIX] sign: better sign item's text sizing
When opening a document to sign (to edit it, not actually sign it) the
sign box fields contains text that are not really readable due to their
funky sizing (too big and truncated).

This commit fixes it by making the font size relative to the one of the
sign box and making it scale correctly. Also, it better aligns the
move-handler and the responsible field.

opw-2459952

closes odoo/enterprise#16517

X-original-commit: ca60f82d1d51caf2ef468ff2ec88f1d5fd0a8a23
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>