[FIX] web_studio: report translations in translations menu
before this commit: there is no easy way to edit report translations in the
studio, the "Translations" menu only show fields and model views, but doesn't
show the model reports even if they are editable with the studio.

after this commit: we display translations related to reports in "Translations"
menu.

note: for a non-stable version, the translation of report should be
improved instead of just adding everything inside "Translations" menu.

opw-2440599
task-2456107

closes odoo/enterprise#16942

X-original-commit: f95bd13b2fad8ac827090ae95ffe4b9f4798983b
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Co-authored-by: Mohammed Shekha <msh@odoo.com>
Co-authored-by: Nicolas Lempereur <nle@odoo.com>