[FIX] web_enterprise: multi-selection on tablet (touch and not small)
On iPad Safari, when using list view, Odoo have trouble to register
clicks from external mouse/trackpad. And when using my fingers to
operate, I cannot bulk edit fields on the list view (e.g. change
salesperson, add tags) because taping on the column will result in
check/uncheck of the record instead of editing the field.

This commit fixes the condition to enable the "mobile-friendly" behavior
on list views.

Note: by removing entirely the "touch" part of the condition, we assume
that mobile-like devices (aka smartphone-like screen size) are also
touch-enabled. It matches the rest of the codebase and was probably a
bit too restrictive in the first place...

Task ID: 2446829

closes odoo/enterprise#16568

X-original-commit: 6fe4a713ea0117c09c29008d4b7bee97a676dbdb
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>