[FIX] helpdesk_sale_timesheet: display SOL in ticket even no timesheets
Before this commit, the SOL in the helpdesk ticket is visible only if
the use_helpdesk_sale_timesheet=True and the partner is defined and at
least a timesheet is defined in this ticket.

The last condition has been removed because we added it when the
sale_line_id is not stored in this model but is no longer the case now.

This commit removes this last condition since this condition is no
longer usefull and now this field will be visible is the
use_helpdesk_sale_timesheet=True and the partner is defined in the
ticket.

closes odoo/enterprise#16616

X-original-commit: 2e9afa1ad7e753bc4d5ff97f28b74841414d499a
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Xavier <xavierbol@users.noreply.github.com>