[FIX] accounting follow-up: avoid wrong follow-up status due to incorrect "followup_line_id"
In follow-up reports with no negative follow-up levels, in certain circumstances you can have a partner with 0$
overdue and still appearing as "In need of action". This can lead to a partner receiving multiple mails while
the actually have no overdue invoices.

Before some move lines can get a "followup_line_id" beside the fact that they are not overdue because the
parameter "aml.date" is taken into account. The parameter "aml.date" should only be compared if there is
no "aml.date_maturity" when we check if the invoice is overdue.

opw-2370615

closes odoo/enterprise#16590

X-original-commit: 437f9abcb385f361eaf4e276e1123504e34d6f1e
Signed-off-by: William André (wan) <wan@odoo.com>