[IMP] mrp_workorder: adapt test
The "quantity_done" field is filled in by onchange. In previous test,
this number is always 0 because we didn't trigger the onchange. Since
now when "quantity_done" is 0 and no backorder, the move will be
cancelled, the test will fail. we adapt the test to trigger the onchange.

Task 2446915
COM PR odoo/odoo#66583
PR #16554