[IMP] project_enterprise: compute default planned dates for task gantt view
Before this commit, when the user create a task via the task gantt view,
the begin and end dates are not based to the calendar of the user or the
one of the current company.

This commit computes the default planned dates to fit with the working
schedule of the user in the task or with the one in the current user or
with the one in the current company.

task-2387984

Community PR: odoo/odoo#66740
Upggrade PR: odoo/upgrade#2203

closes #16627