[IMP] industry_fsm_sale: update the visibility condition of create SO
Before this commit, since the pricing_type field has been changed and
becomes a compute non stored field. The "Create Sales Order" button is
also visible on fsm projects but it should not be the case.

This commit updates the visibility condition of this button to hide this
button when it is a fsm project.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164
closes #16356