[FIX] hr_payroll_account: Make method more heritable
It is currently impossible to modify the creation of the accounting
entry lines.

This commit allows at least to replace the new function
action_create_account_move without altering the behavior of
action_payslip_done. To avoid many differences, it is proposed this way,
but it could segregate more functions.

closes odoo/enterprise#16659

closes odoo/enterprise#30380

X-original-commit: 6e3c35a112393921600cbb6efa56531f860e1833
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>