[IMP] hr_payroll: Auto-update some data file on version increase
Purpose
=======

It happens every months that salary rules and hr_rule_parameters are adapted
and requires a whole module update, trigger a cascading update on all the
dependencies.

Add the possibility to reload some data files each time that the module
version is increased.

This allow to transparently update the parameters on source code
update, every day during the night.

TaskID: 2422704