[FIX] website_sale_account_taxcloud, *: cannot edit an erroneous address
*: website_sale_taxcloud_delivery

Steps:
- Make sure your company is in the United States
- Install eCommerce.
- Set up TaxCloud credentials
  Follow this doc to be sure:
  https://www.odoo.com/documentation/user/14.0/accounting/fiscality/taxes/taxcloud.html
- Go to Invoicing > Configuration > Fiscal Positions
- Edit Fiscal Position for United States:
  - Detect Automatically: checked
- Go to Customers > Products
- Edit a Product (1):
  - TaxCloud category: Select one
- Log out
- Go on the website shop
- Add (1) to the cart
- Checkout the cart
- Enter an address:
  - Zip: 49105
  - State: Massachusetts
- Click Next
- Come back to the erroneous address
- Try to edit it

Bug:
An error is displayed, even if the new address is valid.

Explanation:
The check is done before updating the address. Therefore, the old
address is still used.
Also, when first entering the erroneous address, the validation is not
done since the fiscal position is still unknown:
https://github.com/odoo/enterprise/blob/dc1a1485b164e2ae52aa63ef64dbafd8eedecdc4/website_sale_account_taxcloud/controllers/main.py#L15

This commit moves the check after the creation of the address, during
the payment process. This makes it possible to edit the address
afterwards.
Also, because `website_sale_taxcloud_delivery` depends on
`website_sale_account_taxcloud`, we can remove the override of
`payment()` and let the dependencies do their job:
https://github.com/odoo/enterprise/blob/8b9ebbc2b23bd2f6d8c2bd855b7e03e86ef5f6bb/website_sale_taxcloud_delivery/__manifest__.py#L8
Since https://github.com/odoo/enterprise/pull/2282/commits/4eac2203ef29d6218cfeee3507bf1e0f54dc9384
seems to be a copy of https://github.com/odoo/enterprise/commit/f6836316c0c4b110b61672610a6d24765b415257
we can assume that a second override is not needed.

This commit now lets the user enter any address, but prevents them from
paying if the address is not a valid one according to TaxCloud.

opw:2466467

closes odoo/enterprise#16734

X-original-commit: a56f68d94d8c49f97727010e548c23187c5ca6de
Signed-off-by: backspac <backspac@users.noreply.github.com>
Co-authored-by: Jorge Pinna Puissant <jpp@odoo.com>