[FIX] sale_subscription: fix the tool-tip of the field "payment_token_id"
steps to follow :
Go to Subscription > Create a new subscription :
in the form view go to other info > payment token - tool tip

problem :
the feature is correct, If you didn't set the payment token, it won't charge any card, even if credit cards are set on the partner. but The current tool-tip is wrong, it indicates "If not set, the default payment token of the partner will be used".

solution :

edit tool-tip to: "If not set, the automatic payment will fail".

opw-2463821

closes odoo/enterprise#16701

X-original-commit: cecf0c7fa5118851c4969ee13139fe655b407622
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Signed-off-by: Djamel Touati <DjamelTouati@users.noreply.github.com>