[FIX] industry_fsm: use calendar/gantt specific date
With the change in 87c0ef3 (task-2303945) the date selected by the
calendar or gantt view when creating a new task would no longer be taken
into account.

Since the intent was to avoid the use case where the task was added for
wrong range (eg. clicking on "Add" on month gantt view would add it for
the whole month) this changeset replicate the code only if the default
received range is over one day (computed as over 25 hours for DST).

opw-2442234

closes odoo/enterprise#16719

X-original-commit: e729496a30ec937208ce726d1ec1a2eb14391488
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>