[FIX] account_bank_statement_import_camt: check amount currency
Let USD be the company's currency. When importing a CAMT file, if the
latter contains amounts in another currency, the module may use these
amounts in USD.

To reproduce the error, import this file:
[2449635_CAMT_example.xml](https://drive.google.com/file/d/1t_z9fJLGJsCIsI4i7DjuxMUvCvn1eiGl/view?usp=sharing)

Once done, the user is redirected to the reconciliation page and has to
reconcile $250. However, here is an extract of the CAMT bank statement
XML:
```xml
<Ntry>
    <Amt Ccy="USD">500.00</Amt>
    ...
    <AmtDtls>
        <TxAmt>
            <Amt Ccy="EUR">250.00</Amt>
        </TxAmt>
        <CntrValAmt>
            <Amt Ccy="USD">500.00000</Amt>
            <CcyXchg>
                <SrcCcy>EUR</SrcCcy>
                <TrgtCcy>USD</TrgtCcy>
                <XchgRate>2</XchgRate>
            </CcyXchg>
        </CntrValAmt>
    </AmtDtls>
    <NtryDtls>
        <TxDtls>
            <Amt Ccy="EUR">250.00000</Amt>
            <CdtDbtInd>CRDT</CdtDbtInd>
        </TxDtls>
    </NtryDtls>
</Ntry>
```
Two errors happens:
- The module selects 250 instead of 500
- It displays $250, but this amount is in EURO

The fix ensures that the selected amount is in the same currency than
the company's one. Moreover, the fix also reads the `CntrValAmt` node.

OPW-2449635

closes odoo/enterprise#16916

X-original-commit: 9ba2fcb627891d8a13bd8f4bec2dcc06d074877c
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>