[IMP] helpdesk_sale_timesheet: not filter SOLs if SO defined in project
This commit removes the filter on the SOL on ticket and timesheets on
the ticket if the SO in the project is defined.
However, for the moment, the SO in the project linked to a helpdesk team
should not be set because the pricing_type on these projects must only
be task rate. Thus, the filter is useless for the helpdesk.ticket.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164

closes #16356