[FIX] planning: don't use previous date to recalculate the end datetime
Before this commit, when the user creates a planning slot template
with:
 - start_time = 11 pm
 - duration = 3 hours
and he wants to add a planning.slot using this template in the
employee E's planning, the date range of this shift is on the
same date and the end_datetime is less than start_datetime.
Moreover, this problem implies negative allocated hours.

This commit removes the difference done with the previous end date of
the slot and the new computed end date to recalculate correctly this
date with only the template.
We can ignore the defined range of days previously defined because the
compute for the end_datetime (called `_compute_datetime`) in
planning.slot is not no longer triggered by a change in the data of the
employee linked to the planning.slot but only the template (done in
this commit: 4a7ce5cb7 ).

Cause of this bug:
 - Commit: 87bd8a5

Impacted versions:

 - 14.0 and above versions

Steps to reproduce:

 1. Create planning.slot.template with these information:
    - start_time: 11 pm
    - duration: 3 hours
 2. Go back to the gantt view of planning and click on the '+' button
when we hover a day in the gantt view.
 3. Create this planning.slot using the template created before for a
employee E.

Current behaviour:

 - The date range of this planning.slot is on the same day. For
instance, we have the start_datetime is equal to 1/1/2021 at 11pm
and the end_datetime is equal to 1/1/2021 at 2am.

Expected behaviour:

 - The date range of this planning.slot should be on 2 days and not only
one to have positive allocated hours. If we retake our example above,
the start_datetime should be equal to 1/1/2021 at 11pm and the
end_datetime should be equal to 1/2/2021 at 2am. And then, the allocated
hours should be 3 hours (the duration in the template).

task-2463619

closes odoo/enterprise#16831

X-original-commit: 002b2b744acba3cc2d08c440bc4286c8d1ca9c47
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>