[FIX] account_reports: make ids of hierachy lines unique
Before that, the hierarchy option was using directly ids of account.group objects with a 'hierarchy_' prefix to generate its lines. It wasn't correct as those hierarchy lines could then share ids (in case an account.group was used multiple times). When folding/unfolding one of those lines, all the others were impacted as well.

This is typically the case why we introduced the generic id mechanism, so we just fix the issue by using it.

closes odoo/enterprise#16718

X-original-commit: f2880448fd8c4ebc18b16e92ff0261b6a6537ba2
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>