[IMP] industry_fsm_sale: convert SO field in project into related field
In community, the sale_order_id field has been converted into a related
field. Now the sale_order_id is the sales order of the default sales
order item defined in the project.

This commit removes the override made in the project form view to update
the attributes of the sale_order_id because this field is no longer
visible in the project form view. It also update a unit test to remove
the sale_order_id used in write method, and finally, remove the compute
method for this field, because this compute has been removed in
sale_timesheet module.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164

closes #16356