[IMP] helpdesk_sale_timesheet: find SOL when the partner is different
For the moment, the pricing_type of helpdesk project is either False or
task_rate, but if we want to add the other pricing type in the future
and the user wants to define another partner in the ticket. As in the
task we want to search the last prepaid service SOL in which the
remanining hours is greater than 0.

This commit checks if the partner is the same than the one in the
project to add an additional condition in the domain to search the last
prepaid service SOL.

task-2442683

Community PR: odoo/odoo#66181
Upgrade PR: odoo/upgrade#2164

closes #16356