[IMP] helpdesk: auto close the tickets
Purpose of the task is, If often happens that a customer stops
answering to his ticket when he feels that he was provided a good
solution to this problem. In such case, the support agent has to
manually scan of all the tickets he is in charge of, chase the
customer and move the ticket to a closing stage.
Support agents would save a lot of time if this process was automated.

so in this commit, added the following setting under helpdesk team:
 - name: Automatic Closing
 - description: Close inactive tickets automatically
    - 'Inactivity period _____ days' field with default value = 7 days
    - 'In Stages' m2m, not required; should take all the stages of the
team into account if left empty, and be restricted to the selected stages if set
    - 'Move to Stage' m2o required
when a ticket reaches the 'inactivity period' defined (without interruption) in
one of the selected 'in stages' stages, move the ticket automatically to the
'move to stage' stage
 - also rename the 'ticket closing' feature into 'Ticket Self-closing'

closes odoo/enterprise#16828

Taskid: 2466235
Related: odoo/upgrade#2230
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: IT-Ideas <ltu@odoo.com>