[IMP] sign: allow to set the alignment on text field in sign
Currently, When we sign a document, we directly align to center the
text. It's Ok for some contract, but not for all. So we need to
propose to the user to choose the text Alignment.

So in this commit, added the option to allow the use to set the
alignment on text.

closes odoo/enterprise#17247

Taskid: 2374659
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>