[IMP] sale_timesheet_enterprise: move two settings in project
Before this commit, the allow_billable can be before the
allow_timesheet_timer if the user installs sale_timesheet before
timesheet_grid. But we want to always keep the
allow_timesheet_timer option near allow_timesheets option and
allow_billable option with allow_materials option.

This commit creates a new view in sale_timesheet_enterprise module
because this module need the both modules (sale_timesheet and
timesheet_grid) where there is this problem to be installed.
This new view inherits the same view of the one in the both modules and
its goal is to reorganise these settings in the project form view.

task-2387984

Community PR: odoo/odoo#66740
Upgrade PR: odoo/upgrade#2203

closes #16627