[IMP] social_push_notifications,website_event_social: consolidate visitors push tokens
PURPOSE

Allow consolidation of push tokens on visitors and its children.

SPECS

This commit changes the way push tokens are registered onto website.visitors.
Since the refactoring of events and event.tracks, it is now possible to send
push notifications to users having wishlisted a specific track.

We introduced the concept of parent visitors and children visitors, allowing
the following use case:
- You connect on device (PC) 1 on your Odoo Account 'Michel'
- You enable push notifications and wishlist track 'Amazing OXP showcase'
- You then connect on device 2 on your Odoo Account 'Michel'
- The visitor from device 1 becomes a child of the one from device 2
- You enable push notifications on device 2 as well
  -> Ideally, when the 'Amazing OXP showcase' track is about to start, we would
  like to notify both devices, which was made possible by notifying both your
  parent visitor on device 2 as well as its children (here device 1).

Because we were working on the stable branch, this concept of parent/children
was introduced in new event modules, as opposed to the website module.

This lead to a few issues:
- Complicated code in the website_event_track_social module to determine the
  wishlisted tracks of a visitor, we had to take both the tracks wishlisted by
  that record and also the ones wishlisted by its parent ;
- Having to account for archived visitors when sending push notifications to
  event attendees, since children records are automatically archived ;
- Misleading record count when sending a push notification to the event
  attendees, since it would count both the parent and the children visitors.

The new implementation is a cleaner and also makes a lot of sense with other
visitors implementations:
- A visitor can have multiple leads, all regrouped on its main visitor record ;
- A visitor can have multiple wishlisted tracks, all regrouped on its main
  visitor record ;
- And now a visitor can have multiple push subscriptions, also regrouped on its
  main visitor record.

LINKS

Task-2429652

closes odoo/enterprise#16781

Related: odoo/odoo#65113
Related: odoo/upgrade#2111
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>