[FIX] web_map: update the URL for Google Maps redirection with state
Currently, the URL for the 'View on Google Maps' button is set correctly
when the view is loaded. But with state change, the URL is not in sync
with state, and so the redirection on Google Map always displays the
addresses of records from the initial state only.

This commit fixes the behavior by updating the URL with state change and
thus displays addresses of records from the current state on Google Map.

Task Id:2275169

closes odoo/enterprise#16884

X-original-commit: def3f5371a2320e4c7e04ed9588a1865a758feb0
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>