[FIX] web_studio: change a new binary field's attribute
Have an original view like

```xml
<form>
  <field name="dolores" />
</form>
```

Create a new binary field in a form view (1)
Stay on the form without doing anything else
Click on the new binary
Change its widget (and put the PDF viewer for example) (2)

Before this commit, there is a crash, because at step (1)
a filename char field has been put in the view with the same
xpath spec as the binary field itself
The studio view is like this:

```xml
<data>
  <xpath expr="//field[1]" position="after">
    <field name="filename_field"/>
  </xpath>
  <xpath expr="//field[2]" position="after">
    <field name="binary_field"/>
  </xpath>
</data>
```

The attribute operation has the binary_field as its target, so naturally, the target xpath
is `expr="field[3]" position="attributes"`.

Now, all operations are still stored to allow the feature undo/redo.

So when applying the attribute operation changes, adding the field binary is done once again
with its position being `expr="field[1]" position="after"`

Which given that some change already happened, is a `move` operation for the binary_field
So, when the attribute operation wants to applied, there is no node matching the xpath `expr="field[3]" position="attributes"`
Hence a crash

After this commit, we force the filename for the binary file to be after the binary file
and the binary file keeps the anchor it was given the only source of truth.

OPW 2469683

X-original-commit: cbf0f92831aa13b4e35a90d8266c4615711c64a8