[FIX] mrp_workorder: remove default qty_done value
Having a default of qty_done = 1.0 for quality checks didn't make sense
in any case when the qty to do wasn't 1.0. This made it so if a
workorder was started before the work order's components are reserved
then the qty_done would be stuck on 1.0 instead of the reserved quantity
after the components are reserved.

Steps to reproduce:
- create a BOM with a component and a component quantity != 1 (make sure
  the component is in stock)
- add an operation with a "Register Consumed Materials" Step with
  "Product To Register" = the component
- create a MO with the BOM => confirm => click on Tablet view of the
  generated WO
- WO will display Quantity = 1.00 / 0.00 Units
- Return to MO and click "Check Availability" => click on WO Tablet
  view again

Expected result: Quantity = BOM quantity / 0.00
Actual result: Quantity = 1.00 / 0.00

opw-2448484

closes odoo/enterprise#16847

X-original-commit: e99f10463e18f93254c6b5bc5ec4e6c677c70b7a
Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>