[FIX] l10n_cl_edi: do not skip check certicate and timeout
Previously, because of an update of the urllib3, the
certificate of the SII (CL government) was rejected.

However, the certificate seems to be accepted without
any problems now.  So, it is better to apply the regular
checks to avoid security issues.

Also add default timeout for the urllib and zeep.

X-original-commit: 5e15d7fb75d2fdca21d207a9596d9e19616f49a0
Co-authored-by: Olivier Dony <odony@users.noreply.github.com>