[IMP] l10n_be_hr_payroll: Remove commission structure
Purpose
=======

Commission should be integrated in the classic pay.

The purpose of this structure was to declare the commissions separately, but
this wasn't a good idea. It forced us to duplicate some python methods,
to add rules to compute the ajustements as the commissions are supposed to
be added with the classic pay, etc...

It also duplicate the maintained code and thus extend the risk for desynchronization.

For all those reasons, we decided to remove the whole structure.

TaskID: 2452517