[IMP] helpdesk_sale_timesheet: display price_unit in SOL name_get
Before this commit, when many sales order items have a same product with
a different price unit in a same sale order, the user cannot know which
one of these sales order items has the product with the expensive price
unit, for instance, with only the name_get of the SOL in the ticket form
view.

This commit adds a context in the sale_line_id field of ticket and
so_line field in timesheet for the ticket form view to display the price
unit in the name_get of the sales order items only for the ones in a
sales order have the same product.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes #16348