[IMP] account_reports : add control mechanism in reporting
The main goal is to enhance confidence in Odoo reports.
It should help localization define accurate reports that are less subject to configuration issues through the use of a control domain on html.financial.report.line
When a financial report line having this control domain is computed, it is also computed a second time as the sum of the account.move.line having that domain, and if the 2 numbers differ the line will appear in red in the report.
When at least a line is red, a warning is shown at the top with a button that allows the user to run a check to see if an account is appearing in it only once.
And each account of that type is displayed with all its move lines (grouped) on a new tree view.

closes odoo/enterprise#17009

Task: 2452920
Related: odoo/upgrade#2342
Signed-off-by: William André (wan) <wan@odoo.com>