[IMP] test_l10n_be_hr_payroll: Group work entries generation
Purpose
=======

Instead of ensuring the work entries are generated for all the contracts
one by one at the end, re-generate the work entries for all the concerned
contracts, that will be much more efficient.

Result: N times less calls to _get_work_entries_values when computing the
payslips if N is the number of payslips employees.