[IMP] project_enterprise: add unavailable intervals based company calendar
Before this commit, when there is no user in the row of the gantt view,
the unavailable slot is not used based on the working schedule of the
company.

This commit adds the unvailable intervals based on the working schedule
of the company when there is no user in the row.

task-2387984

Community PR: odoo/odoo#66740
Upgrade PR: odoo/upgrade#2203

closes #16627