[FIX] approvals: approval activity can only be generated automatically
Currently it is possible to create an activity without any possibility
to act with. Indeed, the approval activity is only designed to handle
the case when approval is pending, nothing is done if it's not

so do not allow the user to manualy create an apporval activity.
so to hide the 'approval' activity type from activity wizard, make
it archieved.

closes odoo/enterprise#16987

Taskid: 2439015
Related: odoo/upgrade#2263
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>