[FIX] social_facebook: fix a non-determinist bug in the "Format Facebook Message" test
Bug
===
In the test of "Format Facebook Message", we create a social stream.

But during the creation, we will try to fetch the stream posts, and so
which can not be done during the tests.

So we need to mock the social stream model to not fetch the post
on the Facebook API.

closes odoo/enterprise#17003

X-original-commit: 6dd52e40fa80ebacd72b6a1c4ab4bfce8804f08d
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>