[FIX] timesheet_grid: do not handle the Ctrl key
When the user tried to find something on grid view with Ctrl + F
or copy/paste with Ctrl + C/P the timer for the project
with key F or C or P was also started.

So in this commit ensures the project's timer are not
incremented(started) when the Ctrl key is pressed.

closes odoo/enterprise#17292

Taskid: 2462166
X-original-commit: 06463492d95415e875e0c667ad5b29caea0d45ff
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>