[REF] stock_barcode: "OWLify" the application
Replaces the old client actions / widgets by OWL components.
That means the following modules were modified:
  - `stock_barcode`
  - `stock_barcode_mrp_subcontracting`
  - `stock_barcode_picking_batch`
  - `stock_barcode_quality_control`

Notable changes are:

We have now a client side cache, `LazyBarcodeCache`.
With `getRecord`, it will return a record for the asked model and id,
only for record already in cache.
With `getRecordByBarcode`, it will try to find a record for the given
barcode (for the given model, or for any model who uses barcodes). If no
corresping record is found, it will make a RPC to try to retrive the
missing record for this barcode (if one exists), then if a record is
found, it's added to the cache and then returned.

About `stock_barcode_picking_batch`, we had two client action: one for
the batch and one for the creation and configuration of a new batch.
Now, there is only one client action for these two cases.
Also, the user can now open a empty draft batch to configure it (chose
the operation type and add pickings) instead of having an error message
and nothing else.

We try to keep the tests unchanged during the refactor, but some changes
occur for the following reasons:
  - Makes some triggers more specific to avoid non-deterministic fails;
  - Changes some triggers due to the template's updates (e.g.: some css
    class names was changed, some error messages was updated, ...);
  - Some tests could have unnotified errors. As we checked almost every
    tests, we took the opportunity to fix the errors we found;
  - As more parts of the `stock_barcode` app (client side) are now
    asynchronous, some triggers need to be more specific than before to
    be sure all barcode's actions were fully processed before to run
    through the asserts.

The module `stock_barcode_mobile` was merged into `stock_barcode` as it
was in auto-install, required `stock_barcode` and `web_mobile` and as
`stock_barcode` is an enterprise module, the `web_mobile` is already
installed (auto-install and with `web_enterprise` as the only
requirement), which means `stock_barcode_mobile` was always installed
once `stock_barcode` was too.

task-2352413

Co-authored-by: ryv-odoo <ryv@odoo.com>