[FIX] account_intrastat: convert UoM to reference UoM
When generating the intrastat report, if the associated invoice linces
are expressed in UoM different than the reference UoM, the weights and
quantities won't be correct.

(Need stock. Use French Company)
1. In Settings, enable:
    - Intrastat
    - Units of Measure
2. Create a customer C
    - Country: Belgium
    - VAT number: BE0477472701
3. Create a UoM
    - Name: Three
    - Category: Unit
    - Type: Bigger
    - Ratio: 3
4. Create a product P
    - Unit of Measure: Three
    - Weight: 1
5. Create an invoice
    - Customer: C
    - Invoice Line:
        Product: P
        Unit of Measure: Dozens
6. Post the invoice
7. Go to Intrastat Report

Error: The quantity is 1, it should be 12. This value is based on the
quantity of the account move line, but does not consider the UoM. If
this UoM is not the reference unit, the quantity must be converted.
Moreover, there is a similar problem with the weight (also equal to 1 on
the follow-up report).

The condition to use the `uom.factor` is not correct and does not use
the correct operator. When both invoice line and product have the same
UoM category, we can use the factor to convert the invoice line
quantity. It must be a division, not a multiplication (For instance:
'Three' has a factor of 0.33333 => 1 Three / 0.3333 = 3 Units)

The weight must be converted too. `prod.weight` is the weight of 1 X
product, where X is the UoM of the product and is not always the
reference UoM. In the case above, `prod.weight` is 1 kg, but since the
UoM of the product is 'Three', we have: `1 'Three' P-product = 1 kg <=>
3 'Units' P-product = 1kg`. However, the weight is multiplied by the
quantity expressed in `Units`, thus we have to get the weight per
`Units` too (3 'Units' P-product = 1kg <=>  1 'Units' P-product = 0.333
kg)

OPW-2431764

closes odoo/enterprise#17223

X-original-commit: 04ddbed1d2e8070c564bd920c3129fa8a91b61c6
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>