[IMP] industry_fsm_sale: use read_group to count timesheets in tasks
Before this commit, when the user is a Project User and has no access
right for Timesheet App then we have an access error, if he marked as
done a fsm task. It is because we check if there are at least a
timesheet in the task before creating/confirming the SO. This check
search the timesheet of each task in the self one by one.

This commit adds a read_group to count the number of timesheets by task
with one read into the database and use this result to check if the task
has at least one timesheet. If yes, then the process for the task
continue, that is, we create a SO if the task has no SO linked, then we
create the SOLs for the timesheets and finally we confirm the SO.

task-2376382

Community PR: odoo/odoo#66157
Upgrade PR: odoo/upgrade#2160

closes #16348