[IMP] marketing_automation: change domain editor to inline
In Campaigns, when we add new activity in workflow section it opens as a
modal. After that if we want to add activity_filter on that activity
this action also opens another modal on modal, which is a bad UX practice.

This commit changes the behavior of domain editor. It now opens as inline
domain editor instead of having multiple modal layers.

Task ID-2480625

closes odoo/enterprise#17080

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>