[IMP] web_enterprise: show empty fields on mobile
Before this commit, we had empty space on some form views.
It's because the empty fields weren't shown in readonly mode.

But there is no way the user will click on the "edit" button
to fill this field if he doesn't see it...
Plus, this is more consistent with the brand new feature that
allows you to swich to edit mode when you click on a field.

Steps to reproduce:
- Go to Sale
- Open an existing record
- Referrer doesn't appears
=> You have to click on "edit" button to be able to see it

closes odoo/enterprise#17026

Task-id: 1929043
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>