[FIX] sale_amazon_taxcloud: stop synchronizing invoices with TaxCloud
Before this commit, an authorization request would be sent to TaxCloud
when an invoice with a TaxCloud fiscal position was posted, regardless
of whether the invoice was created from an Amazon sales order or not.
This was problematic because Amazon had already included the sale in a
tax report sent to TaxCloud. Hence, the taxes of the sale were reported
twice.

As of this commit, TaxCloud is no longer notified when an invoice
created from an Amazon sales order is posted. Posting regular invoices
with a TaxCloud fiscal position still sends the request to TaxCloud.

opw-2448235

closes odoo/enterprise#17047

X-original-commit: 374b3ff4ca5f839209b9de26be7a69fb80518d2f
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>
Signed-off-by: Antoine Vandevenne (anv) <AntoineVDV@users.noreply.github.com>