[FIX] account_invoice_extract: fix intracom tax prediction not working
As we were clearing the taxes on the invoice line after having set its
description, it would remove any tax set by prediction, most notably,
intracom taxes.

This tax clear was done to prevent using the default tax.
Now, this is done when the invoice line is created, so that it only
removes this default tax, but not the predicted taxes.

closes odoo/enterprise#18346

X-original-commit: 4d546aadee5a11076b0fdbdd1803cb2422a1e86e
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux <lobaudoux@users.noreply.github.com>