[IMP] website_{appointment, *}: add online appointment snippet
*: enterprise

This snippet allows to add a CTA button linking to all or a selection
of appointments, functionally similar to the "share appointment link"
wizard. If a single appointment type is selected, users can also
specify appointment staff users.

Other snippet and options are moved to the snippets directory to match
the structure of `website`.

Task-2574175
Part of #23750
See odoo/odoo#90748

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>