[FIX] l10n_be_hr_contract_salary: Remove constant yearly costs on credit time config
Coming from:
https://github.com/odoo/enterprise/pull/13384
https://github.com/odoo/enterprise/commit/65d148777da25e5f3cee0e50de1d730f1d4dc5cf

Purpose
=======

There are some methods that are defined to keep a constant employer costs when changing
credit time options such as the work time rate or the wage on signature.

One issue could come when a contract is a part time contract with extra time off, and
modifying a field that affects the final yearly cost.

- wage_with_holidays is recomputed, as this is not stored. This field requires the
  final_yearly_costs value to be computed, as there are extra time off, to compute
  the salary deduction due to the time off.
- In the case of credit time contracts, the final yearly costs, which is supposed to
  be a FTE costs, uses the time_credit_full_time_wage to compute its value.
- The time_credit_full_time_wage field is computed from the wage value, and adapted
  according to the work time rate. However, in case of extra-time off contracts, it
  uses the wage_with_holidays, which is being computed in the current transaction,
  and so its value is still 0.

The result is a totally wrong wage_with_holidays. It implies a recomputation of the
final yearly costs, the wage, the full time credit time wage, etc...

Unfortunately this is not an issue that is fixable in a stable release.

So in this commit we:
- Make final_yearly_costs the REAL costs, like it was before.
- Add a new final_yearly_costs_fte field, which is independant, and doesn't
  impact the wage/wage_with_holidays/...

Also, remove all the non stored fields from the _get_advantage_fields result
to avoid recomputing the final yearly costs when you only modify a contract name,
for example.

closes odoo/enterprise#17078

Taskid: 2483016
X-original-commit: 770d7b3c6263bf7671b67caac7506df8d02cda03
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>